# openapi/features/users/user.yaml
openapi: 3.0.3
info:
  title: User API
  version: 1.0.0
  description: Endpoints for user management (create, update, register, set password, search) and roles retrieval.
servers:
  - url: /api/v1/optifact
paths:
  /users:
    post:
      operationId: createUser
      summary: Create a new user
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserApiRequestView'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiResponseView'
        '409':
          description: Email already exists
    get:
      operationId: searchUsers
      summary: Search users with pagination and sorting
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
        - $ref: '../../parameters/common.yaml#/pageParam'
        - $ref: '../../parameters/common.yaml#/sizeParam'
        - $ref: '../../parameters/common.yaml#/sortFieldParam'
        - $ref: '../../parameters/common.yaml#/sortDirectionParam'
        - name: nameFilter
          in: query
          description: Optional name filter (case-insensitive, partial match)
          required: false
          schema:
            type: string
        - name: emailFilter
          in: query
          description: Optional email filter (case-insensitive, partial match)
          required: false
          schema:
            type: string
        - name: phoneFilter
          in: query
          description: Optional phone filter (case-insensitive, partial match)
          required: false
          schema:
            type: string
        - name: roleNameFilter
          in: query
          description: Optional roleName filter (case-insensitive, partial match)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paged list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageApiResponseView'
  /users/{id}:
    get:
      operationId: findUserById
      summary: Find user by id
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiResponseView'
        '404':
          description: User not found
    put:
      operationId: updateUser
      summary: Update an existing user
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserApiRequestView'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiResponseView'
        '404':
          description: User not found
  /users/auth/{id}/password/{code}:
    post:
      operationId: createUserPassword
      summary: Create or set the user's password
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserPasswordRequestView'
      responses:
        '204':
          description: Password created
          content:
            application/json:
              schema:
                $ref: '../../common/schemas.yaml#/ApiSuccessResponse'
        '404':
          description: User not found
  /users/auth/{id}/activate/{code}:
    post:
      operationId: activateUserAccount
      summary: Activate a user's account
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Account activated
          content:
            application/json:
              schema:
                $ref: '../../common/schemas.yaml#/ApiSuccessResponse'
        '404':
          description: User not found
        '409':
          description: Account already enabled
        '400':
          description: Invalid or expired activation code
  /users/auth/register:
    post:
      operationId: registerUser
      summary: Register a new user linked to an actor by reference
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserApiRequestView'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '../../common/schemas.yaml#/ApiSuccessResponse'
        '409':
          description: Email already exists or actor reference invalid
  /users/email/{email}:
    get:
      operationId: findUserByEmail
      summary: Find user by email
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiResponseView'
        '404':
          description: User not found
  /users/roles:
    get:
      operationId: findRoles
      summary: List all roles
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePageApiResponseView'
  /users/auth/login:
    post:
      operationId: login
      summary: Authenticate user and obtain tokens (supports 2FA)
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginApiRequestView'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginApiResponseView'
        '401':
          description: Invalid credentials
  /users/auth/mfa/verify:
    post:
      operationId: verifyMfaCode
      summary: Verify MFA code and finalize login
      parameters:
        - $ref: '../../parameters/common.yaml#/accept-language'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyMfaCodeApiRequestView'
      responses:
        '200':
          description: MFA verified. Returns access and refresh tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginApiResponseView'
        '400':
          description: Invalid or expired MFA code
        '404':
          description: User not found
components:
  schemas:
    LoginApiRequestView:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginPendingResponseView:
      type: object
      required:
        - email
        - mfaEnable
      properties:
        email:
          type: string
          format: email
        mfaEnable:
          type: boolean
    LoginSuccessResponseView:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    LoginResponseView:
      oneOf:
        - $ref: '#/components/schemas/LoginSuccessResponseView'
        - $ref: '#/components/schemas/LoginPendingResponseView'
    PrivilegeView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required: [id, name]
    VerifyMfaCodeApiRequestView:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
    RoleView:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        privileges:
          type: array
          items:
            $ref: '#/components/schemas/PrivilegeView'
    UserResponseView:
      type: object
      required: [id, lastname, email, accountEnabled, accountLocked, accountExpired, credentialExpired, mfaEnabled, totalLoginAttempt, roles]
      properties:
        id:
          type: string
          format: uuid
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        accountEnabled:
          type: boolean
        accountLocked:
          type: boolean
        accountExpired:
          type: boolean
        credentialExpired:
          type: boolean
        mfaEnabled:
          type: boolean
        totalLoginAttempt:
          type: integer
        actorId:
          type: string
          format: uuid
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleView'
    UserApiRequestView:
      type: object
      required: [lastname, email, roleIds]
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        actorId:
          type: string
          format: uuid
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleView'
    RegisterUserApiRequestView:
      type: object
      required: [lastname, email, phone]
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        taxNumber:
          type: string
          description: Tax number used to link the user to an actor
        registrationNumber:
          type: string
          description: Registration number used to link the user to an actor
        actorName:
          type: string
          description: Actor name used to link the user to an actor
    CreateUserPasswordRequestView:
      type: object
      required: [password]
      properties:
        password:
          type: string
          format: password
        confirmPassword:
          type: string
          format: password
    UserApiResponseView:
      allOf:
        - $ref: '../../common/schemas.yaml#/ApiSuccessResponse'
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/UserResponseView'
    RolePageResponseView:
      allOf:
        - $ref: '../../common/schemas.yaml#/PaginationResponseView'
        - type: object
          required:
            - elements
          properties:
            elements:
              type: array
              description: The list of roles
              items:
                $ref: '#/components/schemas/RoleView'
    RolePageApiResponseView:
      allOf:
        - $ref: '../../common/schemas.yaml#/ApiSuccessResponse'
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/RolePageResponseView'
    UserPageResponseView:
      allOf:
        - $ref: '../../common/schemas.yaml#/PaginationResponseView'
        - type: object
          required:
            - elements
          properties:
            elements:
              type: array
              description: The list of users
              items:
                $ref: '#/components/schemas/UserResponseView'
    UserPageApiResponseView:
      allOf:
        - $ref: '../../common/schemas.yaml#/ApiSuccessResponse'
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/UserPageResponseView'
    LoginApiResponseView:
      allOf:
        - $ref: '../../common/schemas.yaml#/ApiSuccessResponse'
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/LoginResponseView'